+++ ContraGame/ContraGame/CEnemy.cpp
@@ -16,7 +16,7 @@ void CEnemy::LoadResources()
 
 }
 
-void CEnemy::Update(float delta_time)
+void CEnemy::Update(int delta_time)
 {
 
 }
