+++ ContraGame/ContraGame/CFBullet.h
@@ -18,7 +18,7 @@ public:
 
 	void LoadResources();
 	void Draw();
-	void Update(float delta_time);
+	void Update(int delta_time);
 	void Moving(float v_max);
 	void Shoot(D3DXVECTOR3 pos, int angle, float v_max, float vo = 0);
 };
