+++ ContraGame/ContraGame/CNBullet.h
@@ -18,7 +18,7 @@ public:
 	~CNBullet();
 
 	void LoadResources();
-	void Update(float delta_time);
+	void Update(int delta_time);
 	void Draw();
 
 	void Moving(float v_max);
